refactor process and bump version
