+++ stormdb/process/__init__.py
@@ -1,0 +1,9 @@
+"""Tools for processing data on hyades-cluster."""
+
+# Authors:  Christopher Bailey <cjb@cfin.au.dk>
+#           Mads Jensen <mads@cfin.au.dk>
+#
+# License: BSD (3-clause)
+
+from .maxfilter import Maxfilter
+from .mne_python import MNEPython
