+++ stormdb/process/base.py
@@ -1,0 +1,27 @@
+"""
+=========================
+Helpers for process-tools.
+=========================
+"""
+
+import os
+
+
+def check_destination_writable(dest):
+    try:
+        open(dest, 'w')
+    except IOError:
+        return False
+    else:
+        os.remove(dest)
+        return True
+
+
+def check_source_readable(source):
+    try:
+        fid = open(source, 'r')
+    except IOError:
+        return False
+    else:
+        fid.close()
+        return True
